hrm/project/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-rest-coreapi-schema==0.1.2
djangorestframework==3.7.7
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==2.6.10
MarkupSafe==2.0.1
mysqlclient==2.0.3
Pygments==2.2.0
python-decouple==3.5
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
